requirements.txt
numpy
pandas
PyQt6
PyQt6-Qt6
PyQt6-sip
pyqtgraph
pyserial
python-dateutil